feat: complete add whitelist for resource
