+++ apps/builder/src/components/WhiteList/index.tsx
@@ -1,0 +1,108 @@
+import { AxiosResponse } from "axios"
+import copy from "copy-to-clipboard"
+import { useCallback, useEffect, useState } from "react"
+import { useTranslation } from "react-i18next"
+import {
+  Card,
+  CopyIcon,
+  DownIcon,
+  UpIcon,
+  useMessage,
+} from "@illa-design/react"
+import { Api } from "@/api/base"
+import {
+  IpListContainerStyle,
+  IpListStyle,
+} from "@/page/App/components/Actions/ActionGenerator/ActionTypeSelector/style"
+import { isCloudVersion } from "@/utils/typeHelper"
+import {
+  whiteListButtonContainerStyle,
+  whiteListButtonStyle,
+  whiteListContentContainerStyle,
+  whiteListContentStyle,
+  whiteListDescriptionStyle,
+  whiteListOperationIconStyle,
+  whiteListTitleStyle,
+} from "./style"
+
+export const WhiteList = () => {
+  const { t } = useTranslation()
+  const message = useMessage()
+
+  const [showIPList, setShowIPList] = useState<boolean>(false)
+  const [ipList, setIPList] = useState<string[]>([])
+
+  useEffect(() => {
+    if (!isCloudVersion) {
+      return
+    }
+    Api.request(
+      {
+        url: "https://peripheral-api.illasoft.com/v1/meta",
+        method: "GET",
+      },
+      (data: AxiosResponse<{ resources: string[] }>) => {
+        const { resources } = data.data
+        setIPList(resources)
+      },
+    )
+  }, [])
+
+  const handleOperationIconClick = () => {
+    setShowIPList((prevState) => !prevState)
+  }
+
+  const handleCopyClick = useCallback(() => {
+    const copyResult = copy(ipList.join("\n"))
+    if (copyResult) {
+      message.success({
+        content: t("copied"),
+      })
+    } else {
+      message.error({
+        content: t("copy_failed"),
+      })
+    }
+  }, [ipList, message, t])
+
+  return (
+    <div>
+      <div css={whiteListContentContainerStyle}>
+        <div css={whiteListContentStyle}>
+          <div css={whiteListTitleStyle}>
+            {t("editor.action.resource.tip.allowlist.title")}
+          </div>
+          <div css={whiteListDescriptionStyle}>
+            {t("editor.action.resource.tip.allowlist.message")}
+          </div>
+        </div>
+        {isCloudVersion && (
+          <div css={whiteListButtonContainerStyle}>
+            <div css={whiteListButtonStyle} onClick={handleCopyClick}>
+              <CopyIcon />
+              <span>{t("editor.action.resource.button.copy_ip")}</span>
+            </div>
+            <span
+              css={whiteListOperationIconStyle}
+              onClick={handleOperationIconClick}
+            >
+              {showIPList ? <UpIcon /> : <DownIcon />}
+            </span>
+          </div>
+        )}
+      </div>
+      {isCloudVersion && showIPList && (
+        <div css={IpListContainerStyle}>
+          <Card>
+            <div css={IpListStyle}>
+              {ipList.map((ip) => (
+                <div key={ip}>{ip}</div>
+              ))}
+            </div>
+          </Card>
+        </div>
+      )}
+    </div>
+  )
+}
+WhiteList.displayName = "WhiteList"
