+++ apps/builder/src/page/Dashboard/components/ResourceGenerator/ResourceTypeSelector/index.tsx
@@ -1,46 +1,12 @@
-import copy from "copy-to-clipboard"
-import { FC, useCallback, useState } from "react"
-import { useTranslation } from "react-i18next"
-import { CopyIcon, DownIcon, UpIcon, createMessage } from "@illa-design/react"
+import { FC } from "react"
+import { WhiteList } from "@/components/WhiteList"
 import { ResourceCard } from "@/page/Dashboard/components/ResourceGenerator/ResourceCard"
 import { ResourceTypeList } from "@/page/Dashboard/components/ResourceGenerator/config"
-import { isCloudVersion } from "@/utils/typeHelper"
 import { ResourceTypeSelectorProps } from "./interface"
-import {
-  categoryStyle,
-  containerStyle,
-  resourceListStyle,
-  whiteListButtonContainerStyle,
-  whiteListButtonStyle,
-  whiteListContentContainerStyle,
-  whiteListContentStyle,
-  whiteListDescriptionStyle,
-  whiteListOperationIconStyle,
-  whiteListTitleStyle,
-} from "./style"
+import { categoryStyle, containerStyle, resourceListStyle } from "./style"
 
 export const ResourceTypeSelector: FC<ResourceTypeSelectorProps> = (props) => {
   const { onSelect } = props
-  const { t } = useTranslation()
-  const message = createMessage()
-
-  const [showIPList, setShowIPList] = useState<boolean>(true)
-
-  const handleOperationIconClick = () => {
-    setShowIPList((prevState) => !prevState)
-  }
-  const handleCopyClick = useCallback(() => {
-    const copyResult = copy("")
-    if (copyResult) {
-      message.success({
-        content: t("copied"),
-      })
-    } else {
-      message.error({
-        content: t("copy_failed"),
-      })
-    }
-  }, [])
 
   return (
     <div css={containerStyle}>
@@ -60,33 +26,7 @@ export const ResourceTypeSelector: FC<ResourceTypeSelectorProps> = (props) => {
           </div>
         </div>
       ))}
-      <div>
-        <div css={whiteListContentContainerStyle}>
-          <div css={whiteListContentStyle}>
-            <div css={whiteListTitleStyle}>
-              {t("editor.action.resource.tip.allowlist.title")}
-            </div>
-            <div css={whiteListDescriptionStyle}>
-              {t("editor.action.resource.tip.allowlist.message")}
-            </div>
-          </div>
-          {isCloudVersion && (
-            <div css={whiteListButtonContainerStyle}>
-              <div css={whiteListButtonStyle} onClick={handleCopyClick}>
-                <CopyIcon />
-                <span>{t("editor.action.resource.button.copy_ip")}</span>
-              </div>
-              <span
-                css={whiteListOperationIconStyle}
-                onClick={handleOperationIconClick}
-              >
-                {showIPList ? <UpIcon /> : <DownIcon />}
-              </span>
-            </div>
-          )}
-        </div>
-        {isCloudVersion && showIPList && <div></div>}
-      </div>
+      <WhiteList />
     </div>
   )
 }
